Train Brain Vanilla
I did a dumb with this, I made it on the ground with vanilla capabilities, then converted it to Space Exploration and made it in space because it was visually clearer, but then I bugtested further and ended up tweaking the timing. Those tweaks aren't in this vanilla blueprint. You may have to mess around with the settings on the trains a bit, mostly when they reset, to get it to work perfectly smoothly. If they lock, try adding five or ten [dot] to the train that locked, as well as every one that goes after it.

Thread on reddit: https://www.reddit.com/r/factorio/comments/12d6mv1/mrw_i_see_someone_making_a_7_segment_display_on/